Template Auth for login and register
