+++ src/views/Auth/Logout/index.js
