+++ src/views/Auth/Register/Register.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import { Title } from '../styles';
+
+class Register extends React.Component {
+  render() {
+    return <Title>Criar uma conta</Title>;
+  }
+}
+
+export default Register;
