+++ src/views/Auth/Register/index.js
@@ -1,0 +1,3 @@
+import Register from './Register';
+
+export default Register;
